\name{FSA_annotation_text_repel}
\alias{FSA_annotation_text_repel}
\title{
FSA annotation text repel
}
\description{
This function is to set annotations on the spectra plots with a reasonable distance to avoid overlying annotations.
}
\usage{
FSA_annotation_text_repel(FSAspectra, nGridX, nGridY)
}
\arguments{
  \item{FSAspectra}{
FSAspectra
}
  \item{nGridX}{
number of grids on the x-axis
}
  \item{nGridY}{
number of grids on the y-axis
}
}
\value{
labels
}