\name{FSA_logRecorder}
\alias{FSA_logRecorder}
\title{
FSA logRecorder
}
\description{
FSA_logRecorder
}
\usage{
FSA_logRecorder(messageQuote, allowedPrinting = TRUE)
}
\arguments{
  \item{messageQuote}{
messageQuote
}
  \item{allowedPrinting}{
allowedPrinting
}
}
\value{
a line of communication messages is exported to the console and the log .txt file.
}