\name{FSA_message}
\alias{FSA_message}
\title{
FSA message
}
\description{
FSA_message
}
\usage{
FSA_message(messageQuote, failedMessage= TRUE)
}
\arguments{
  \item{messageQuote}{
messageQuote
}
  \item{failedMessage}{
failedMessage
}
}
\value{
a line of communication messages is exported to the console.
}