\name{FSA_R.aggregate}
\alias{FSA_R.aggregate}
\title{
aggregate function for IDSL.FSA
}
\description{
This module ensures that the `aggregate` function of R returns a list type of data.
}
\usage{
FSA_R.aggregate(FSAvec)
}
\arguments{
  \item{FSAvec}{
a vector of data
}
}
\value{
listIDFSAvec
}