\name{FSA_workflow}
\alias{FSA_workflow}
\title{
FSA workflow
}
\description{
This function executes the FSA workflow.
}
\usage{
FSA_workflow(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
FSA spreadsheet
}
}
\value{
This function organizes the FSA file processing for better performance using the template spreadsheet.
}