\name{FSA_xlsxAnalyzer}
\alias{FSA_xlsxAnalyzer}
\title{
FSA xlsx Analyzer
}
\description{
This function processes the spreadsheet of the FSA parameters to ensure the parameter inputs are consistent with the requirements of the IDSL.FSA pipeline.
}
\usage{
FSA_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
FSA spreadsheet
}
}
\value{
This function returns the FSA parameters to feed the FSA_workflow function.
}
