\name{FSdb_file_generator}
\alias{FSdb_file_generator}
\title{
FSdb file generator
}
\description{
This function generates FSDB objects
}
\usage{
FSdb_file_generator(PARAM_FSdb, output_path = NULL)
}
\arguments{
  \item{PARAM_FSdb}{
`PARAM_FSdb` parameters obtained by the `FSA_FSdb_xlsxAnalyzer` function.
}
  \item{output_path}{
output_path
}
}
\value{
An FSDB object
}