\name{FSdb_subsetter}
\alias{FSdb_subsetter}
\title{
FSdb subsetter
}
\description{
FSdb subsetter
}
\usage{
FSdb_subsetter(FSdb, inclusionIDs = NULL, exclusionIDs = NULL)
}
\arguments{
  \item{FSdb}{
FSdb
}
  \item{inclusionIDs}{
inclusionIDs
}
  \item{exclusionIDs}{
exclusionIDs
}
}
\value{
subsetted FSdb
}