\name{UFSA_element_sorter}
\alias{UFSA_element_sorter}
\title{
Element Sorter
}
\description{
This function sorts 84 elements in the periodic table for molecular formula deconvolution.
}
\usage{
UFSA_element_sorter()
}
\value{
A string vector of elements
}
\examples{
Elements <- UFSA_element_sorter()
}
